+++ src/transport/smtp/extension.rs
@@ -13,6 +13,7 @@ use std::{
 /// Client identifier, the parameter to `EHLO`
 #[derive(PartialEq, Eq, Clone, Debug)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+#[non_exhaustive]
 pub enum ClientId {
     /// A fully-qualified domain name
     Domain(String),
@@ -57,14 +58,9 @@ impl Display for ClientId {
 }
 
 impl ClientId {
-    #[deprecated(since = "0.10.0", note = "Please use the new_domain function instead")]
+    #[deprecated(since = "0.10.0", note = "Please use ClientId::Domain(domain) instead")]
     /// Creates a new `ClientId` from a fully qualified domain name
     pub fn new(domain: String) -> Self {
-        Self::Domain(domain)
-    }
-
-    /// Creates a new `ClientId` from a fully qualified domain name
-    pub fn new_domain(domain: String) -> Self {
         Self::Domain(domain)
     }
 }
@@ -296,7 +292,7 @@ mod test {
     #[test]
     fn test_clientid_fmt() {
         assert_eq!(
-            format!("{}", ClientId::new_domain("test".to_string())),
+            format!("{}", ClientId::Domain("test".to_string())),
             "test".to_string()
         );
         assert_eq!(format!("{}", LOCALHOST_CLIENT), "[127.0.0.1]".to_string());