+++ src/transport/smtp/mod.rs
@@ -162,7 +162,7 @@
 //! # {
 //! use lettre::transport::smtp::{SMTP_PORT, extension::ClientId, commands::*, client::SmtpConnection};
 //!
-//! let hello = ClientId::new_domain("my_hostname".to_string());
+//! let hello = ClientId::Domain("my_hostname".to_string());
 //! let mut client = SmtpConnection::connect(&("localhost", SMTP_PORT), None, &hello, None).unwrap();
 //! client.command(
 //!         Mail::new(Some("user@example.com".parse().unwrap()), vec![])