+++ CHANGELOG.md
@@ -33,7 +33,7 @@ Several breaking changes were made between 0.9 and 0.10, but changes should be s
 * Merge `Email` and `SendableEmail` into `lettre::Email` ([ce37464](https://github.com/lettre/lettre/commit/ce37464))
 * When the hostname feature is disabled or hostname cannot be fetched, `127.0.0.1` is used instead of `localhost` as
   EHLO parameter (for better RFC compliance and mail server compatibility)
-* The `new` method of `ClientId` is renamed to `new_domain`
+* The `new` method of `ClientId` is deprecated
 
 #### Bug Fixes